--- src/IRoute.php
@@ -1,11 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace wub;
-
-interface IRoute {
-
-	public function getHandlerId(): string;
-
-	public function getParams(): array;
-
-}
