--- src/IRouter.php
@@ -1,9 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace wub;
-
-interface IRouter {
-
-	public function route(string $locator): ?IRoute;
-
-}
