+++ src/Response.php
@@ -1,16 +1,11 @@
 <?php declare(strict_types=1);
 
-namespace wub;
+namespace frm;
 
-/**
- * Ответ
- */
 interface IResponse {
-
 	public function send($file): int;
 
 	public static function notFound(string $message, $context = []): self;
 
 	public static function internalError(string $message, $context = []): self;
-
 }
